Moves custom exceptions to `exceptions.py` module
