+++ main.py
@@ -11,6 +11,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from config import settings
 from database import create_tables, get_db
+from exceptions import SelfReferencingURLError, URLTooLongError, URLTooShortError
 from models import Link
 
 app: FastAPI = FastAPI(title=settings.app_name)
@@ -20,23 +21,6 @@ app: FastAPI = FastAPI(title=settings.app_name)
 async def lifespan(app: FastAPI) -> AsyncGenerator:
     await create_tables()
     yield
-
-""" Custom errors """
-
-
-class URLTooLongError(ValueError):
-    def __init__(self, url: str, limit: int) -> None:
-        super().__init__(f"URL exceeds max length of {limit} characters: {url}")
-
-
-class URLTooShortError(ValueError):
-    def __init__(self, url: str, minimum: int) -> None:
-        super().__init__(f"URL shorter than minimum {minimum} characters: {url}")
-
-
-class SelfReferencingURLError(ValueError):
-    def __init__(self, url: str) -> None:
-        super().__init__(f"Cannot shorten URLs pointing to the service itself: {url}")
 
 
 """Pydantic models"""
